requirements.txt
torch>=2.1.0
transformers>=4.38.0
datasets>=2.17.0
peft>=0.9.0
bitsandbytes>=0.42.0
accelerate>=0.27.0
scipy
pydantic>=2.0.0
pyyaml
fastapi
uvicorn
rich
psutil
tqdm
sentencepiece
protobuf
